// src/app/constants/enums/zb.enum.ts
//     name: '画符剑', 1
//     name: '紫薇剑', 2
//     name: '玄铁剑', 3
//     name: '挽留剑', 4
//     name: '割鹿刀',
//     name: '小楼一夜听春雨',
//     name: '倚天剑',
//     name: '轩辕剑',
//     name: '木棍',
//     name: '烧火棍',
//     name: '打狗棒',
//     name: '阴阳棍',
//     name: '绿玉杖',
//     name: '天机棒',
//     name: '青魔手',
//     name: '玉佛手',
//     name: '细砂',
//     name: '金钱镖',
//     name: '仙鹤神针',
//     name: '情人箭',
//     name: '孔雀翎',
//     name: '小李飞刀',
//     name: '软猬甲',
//     name: '天蚕衣',
//     name: '霓裳羽衣',
//     name: '金缕衣',
//     name: '赤兔',
//     name: '越影',
//     name: '的卢',
//     name: '汗血',
//     name: '奔霄',
//     name: '萧稍',
//     name: '楚骓',
export enum ZbIdEnum {
  None = 0,
  HuafuJian = 1,
  ZiWeiJian = 2,
  XuanTieJian = 3,
  WanLiuJian = 4,
  GeLuDao = 5,
  XiaoLou = 6,
  YiTianJian = 7,
  XuanYuanJian = 8,
  MuGun = 9,
  ShaoHuoGun = 10,
  DaGouBang = 11,
  YinYangGun = 12,
  LvYuZhang = 13,
  TianJiBang = 14,
  QingMoShou = 15,
  YuFoShou = 16,
  XiSha = 17,
  JinQianBiao = 18,
  XianHeShenZhen = 19,
  QingRenJian = 20,
  KongQueLing = 21,
  XiaoLiFeiDao = 22,
  RuanWeiJia = 23,
  TianCanYi = 24,
  NiChangYuYi = 25,
  JinLvYi = 26,
  ChiTu = 27,
  YueYing = 28,
  DiLu = 29,
  HanXue = 30,
  BenXiao = 31,
  XiaoShao = 32,
  ChuZhui = 33,
}